Update pics_and_vids_2.py
Changing the method (i.e., the algorithm) used to hash the images.

The average_hash method was making some images show up with the same hash value, even though these images are not the same. I saw this happen in two scenarios: (1) with the default image that can show up on multiple profiles (this is normal), and (2) with images that have a white bar on the left side of the image and a white bar on the right side of the image (this is NOT normal because it means that different images can have the same hash value).